refactoring reading 2D proj data from DICOM into its own set of .h/.cpp
