+++ apps/image_io/convert_radiograph_dicom_to_proj_data/xreg_convert_radiograph_dcm_to_proj_data_main.cpp
@@ -24,7 +24,7 @@
 
 #include "xregProgOptUtils.h"
 #include "xregH5ProjDataIO.h"
-#include "xregDICOMUtils.h"
+#include "xregReadProjDataFromDICOM.h"
 #include "xregStringUtils.h"
 
 int main(int argc, char* argv[])
