+++ lib/file_formats/xregReadProjDataFromDICOM.cpp
@@ -1,0 +1,55 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2021 Robert Grupp
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+#include "xregReadProjDataFromDICOM.h"
+#include "xregReadProjDataFromDICOMDetail.h"
+
+xreg::ProjDataF32List xreg::ReadProjDataFromDICOMF32(const std::string& dcm_path,
+                                                     const ReadProjDataFromDICOMParams& params)
+{
+  return detail::ReadProjDataFromDICOMHelper<float>(dcm_path, params);
+}
+
+xreg::ProjDataF32List xreg::ReadProjDataFromDICOMF32(const std::string& dcm_path,
+                                                     const std::string& fcsv_path,
+                                                     const ReadProjDataFromDICOMParams& params)
+{
+  return detail::ReadProjDataFromDICOMHelper<float>(dcm_path, fcsv_path, params);
+}
+
+
+xreg::ProjDataU16List xreg::ReadProjDataFromDICOMU16(const std::string& dcm_path,
+                                                     const ReadProjDataFromDICOMParams& params)
+{
+  return detail::ReadProjDataFromDICOMHelper<unsigned short>(dcm_path, params);
+}
+
+xreg::ProjDataU16List xreg::ReadProjDataFromDICOMU16(const std::string& dcm_path,
+                                                     const std::string& fcsv_path,
+                                                     const ReadProjDataFromDICOMParams& params)
+{
+  return detail::ReadProjDataFromDICOMHelper<unsigned short>(dcm_path, fcsv_path, params);
+}
+
+
